Add user can't see edit/delete links spec
